+++ src/main/java/models/TransferRequest.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransferRequest extends BaseModel {
-    private int senderAccountId;
-    private int receiverAccountId;
+    private long senderAccountId;
+    private long receiverAccountId;
     private Double amount;
 
 }
